requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
sentence-transformers==2.2.2
scikit-learn==1.3.2
faiss-cpu==1.7.4
redis==5.0.1
python-dotenv==1.0.0
numpy==1.24.3